refactor: split root and logger
